requirements_jupyter.txt
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
enum-compat==0.0.3
Flask==3.1.2
idna==3.10
intuit-oauth==1.2.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.3.1
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
pandas==2.3.2
openpyxl==3.1.5
numpy==2.3.3

# Jupyter dependencies
jupyter==1.1.1
jupyterlab==4.4.9
jupyter-server==2.17.0
notebook==7.4.7
ipykernel==6.30.1
ipywidgets==8.1.7 